Refactor emissions steps to create new release of CO2 dataset
